refactor(docs): share useAttachmentSrc across the clone examples
five clones carried the same object-url plumbing: three byte-identical, gemini differing only in how it destructured the selector, grok inlining the effect instead of calling the helper. none of it carries any of the visual identity the clones exist to demonstrate.

the demo zip builder collects an explicit file list rather than walking imports, so the shared module is registered alongside the other clone sources; without that the self-containment test fails for all five.
