+++ apps/docs/components/pages/examples/claude.tsx
@@ -28,8 +28,8 @@ import {
   ThumbsUp,
   XIcon,
 } from "lucide-react";
-import { useEffect, useState, type FC } from "react";
-import { useShallow } from "zustand/shallow";
+import { useState, type FC } from "react";
+import { useAttachmentSrc } from "@/components/pages/examples/use-attachment-src";
 import { MarkdownText } from "@/components/assistant-ui/markdown-text";
 import { CloneThreadShell } from "./clone-thread-shell";
 import {
@@ -320,41 +320,6 @@ const ChatMessage: FC = () => {
       </AuiIf>
     </MessagePrimitive.Root>
   );
-};
-
-const useFileSrc = (file: File | undefined) => {
-  const [src, setSrc] = useState<string | undefined>(undefined);
-
-  useEffect(() => {
-    if (!file) {
-      setSrc(undefined);
-      return;
-    }
-
-    const objectUrl = URL.createObjectURL(file);
-    setSrc(objectUrl);
-
-    return () => {
-      URL.revokeObjectURL(objectUrl);
-    };
-  }, [file]);
-
-  return src;
-};
-
-const useAttachmentSrc = () => {
-  const { file, src } = useAuiState(
-    useShallow((s): { file?: File; src?: string } => {
-      if (s.attachment.type !== "image") return {};
-      if (s.attachment.file) return { file: s.attachment.file };
-      const src = s.attachment.content?.filter((c) => c.type === "image")[0]
-        ?.image;
-      if (!src) return {};
-      return { src };
-    }),
-  );
-
-  return useFileSrc(file) ?? src;
 };
 
 const ClaudeAttachment: FC = () => {
