+++ apps/docs/components/pages/examples/gemini.tsx
@@ -28,8 +28,8 @@ import {
   ThumbsUp,
   XIcon,
 } from "lucide-react";
-import { type FC, useEffect, useState } from "react";
-import { useShallow } from "zustand/shallow";
+import { type FC, useState } from "react";
+import { useAttachmentSrc } from "@/components/pages/examples/use-attachment-src";
 import { MarkdownText } from "@/components/assistant-ui/markdown-text";
 import { CloneThreadShell } from "./clone-thread-shell";
 import {
@@ -272,41 +272,6 @@ const ChatMessage: FC = () => {
       </AuiIf>
     </MessagePrimitive.Root>
   );
-};
-
-const useFileSrc = (file: File | undefined) => {
-  const [src, setSrc] = useState<string | undefined>(undefined);
-
-  useEffect(() => {
-    if (!file) {
-      setSrc(undefined);
-      return;
-    }
-
-    const objectUrl = URL.createObjectURL(file);
-    setSrc(objectUrl);
-
-    return () => {
-      URL.revokeObjectURL(objectUrl);
-    };
-  }, [file]);
-
-  return src;
-};
-
-const useAttachmentSrc = () => {
-  const { file, src } = useAuiState(
-    useShallow(({ attachment }): { file?: File; src?: string } => {
-      if (attachment.type !== "image") return {};
-      if (attachment.file) return { file: attachment.file };
-      const src = attachment.content?.filter((c) => c.type === "image")[0]
-        ?.image;
-      if (!src) return {};
-      return { src };
-    }),
-  );
-
-  return useFileSrc(file) ?? src;
 };
 
 const GeminiAttachment: FC = () => {
