+++ apps/docs/components/pages/examples/perplexity.tsx
@@ -31,8 +31,8 @@ import {
   Telescope,
   XIcon,
 } from "lucide-react";
-import { type FC, useEffect, useState } from "react";
-import { useShallow } from "zustand/shallow";
+import { type FC, useState } from "react";
+import { useAttachmentSrc } from "@/components/pages/examples/use-attachment-src";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -383,41 +383,6 @@ const BranchPicker: FC<BranchPickerPrimitive.Root.Props> = ({
       </BranchPickerPrimitive.Next>
     </BranchPickerPrimitive.Root>
   );
-};
-
-const useFileSrc = (file: File | undefined) => {
-  const [src, setSrc] = useState<string | undefined>(undefined);
-
-  useEffect(() => {
-    if (!file) {
-      setSrc(undefined);
-      return;
-    }
-
-    const objectUrl = URL.createObjectURL(file);
-    setSrc(objectUrl);
-
-    return () => {
-      URL.revokeObjectURL(objectUrl);
-    };
-  }, [file]);
-
-  return src;
-};
-
-const useAttachmentSrc = () => {
-  const { file, src } = useAuiState(
-    useShallow((s): { file?: File; src?: string } => {
-      if (s.attachment.type !== "image") return {};
-      if (s.attachment.file) return { file: s.attachment.file };
-      const src = s.attachment.content?.filter((c) => c.type === "image")[0]
-        ?.image;
-      if (!src) return {};
-      return { src };
-    }),
-  );
-
-  return useFileSrc(file) ?? src;
 };
 
 const AttachmentTypeLabel: FC = () => {
